refactor: auction/price

+++ auction/src/main/java/com/smore/auction/application/command/AuctionCreateCommand.java
@@ -1,10 +1,11 @@
 package com.smore.auction.application.command;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record AuctionCreateCommand(
     UUID productId,
-    Long productPrice,
+    BigDecimal productPrice,
     Long stock,
     Long sellerId
 ) {
